+++ src/Routes.js
@@ -3,7 +3,7 @@ import { Redirect, Route, Switch } from 'react-router-dom'
 import Agreement from './components/Agreement/Agreement'
 import Proposals from './components/Proposals/Proposals'
 import ProposalSingle from './components/Proposals/ProposalSingle'
-import Stake from './components/Stake/Stake'
+import StakeManagement from './components/Stake/StakeManagement'
 
 export default function Routes() {
   return (
@@ -12,7 +12,7 @@ export default function Routes() {
       <Route exact path="/proposals" component={Proposals} />
       <Route exact path="/proposals/:id" component={ProposalSingle} />
       <Route exact path="/agreement" component={Agreement} />
-      <Route exact path="/stake-management" component={Stake} />
+      <Route exact path="/stake-management" component={StakeManagement} />
       <Redirect to="/proposals" />
     </Switch>
   )
